Change weight system from float to double for ultra-precise drop rates
- Updated WeightedStack to use double instead of float for weight precision
- Modified MinerRecipe serialization to handle double values
- Updated ListUtil.getTotalWeight() to return double
- Fixed ControllerBaseBE.getWeightedItem() to use double arithmetic
- Updated JEI MinerCategory.customFormat() for double support
- Changed localization strings from %d to %s for decimal weight display
- Updated ModRecipeProvider with double values instead of float literals
- Added ultra-rare Ultimate Stellar Core drop (0.000005 weight) for END dimension

This allows for extremely precise drop rates like 0.000001 (0.0001%)

+++ src/main/java/com/leo/voidminers/block/entity/ControllerBaseBE.java
@@ -604,9 +604,9 @@ public class ControllerBaseBE extends BlockEntity {
     }
 
     public ItemStack getWeightedItem(List<WeightedStack> items, RandomSource random) {
-        float totalWeight = ListUtil.getTotalWeight(items);
+        double totalWeight = ListUtil.getTotalWeight(items);
 
-        float randomValue = random.nextFloat() * totalWeight;
+        double randomValue = random.nextDouble() * totalWeight;
 
         for (WeightedStack item : items) {
             randomValue -= item.weight;
